fix a bug
