+++ mmseg/core/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .evaluation import *  # noqa: F401, F403
 from .layer_decay_optimizer_constructor import \
-    LayerDecayOptimizerConstructor  # noqa: F401, F403
-from .resize_transform import SETR_Resize  # noqa: F401, F403
+    LayerDecayOptimizerConstructor  # noqa: F401
 from .seg import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
