+++ mmseg/models/backbones/vit.py
@@ -2,17 +2,13 @@
 import math
 import warnings
 
-import numpy as np
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 from mmcv.cnn import build_norm_layer
-from mmcv.cnn.bricks.drop import build_dropout
 from mmcv.cnn.bricks.transformer import FFN, MultiheadAttention
 from mmcv.cnn.utils.weight_init import (constant_init, kaiming_init,
                                         trunc_normal_)
 from mmcv.runner import BaseModule, ModuleList, _load_checkpoint
-from scipy import interpolate
 from torch.nn.modules.batchnorm import _BatchNorm
 from torch.nn.modules.utils import _pair as to_2tuple
 
@@ -20,138 +16,6 @@ from mmseg.ops import resize
 from mmseg.utils import get_root_logger
 from ..builder import BACKBONES
 from ..utils import PatchEmbed
-
-
-class BEiTAttention(BaseModule):
-    """Window based multi-head self-attention (W-MSA) module with relative
-    position bias.
-
-    Args:
-        embed_dims (int): Number of input channels.
-        num_heads (int): Number of attention heads.
-        window_size (tuple[int]): The height and width of the window.
-        qkv_bias (bool, optional):  If True, add a learnable bias to q, k, v.
-            Default: True.
-        qk_scale (float | None, optional): Override default qk scale of
-            head_dim ** -0.5 if set. Default: None.
-        attn_drop_rate (float, optional): Dropout ratio of attention weight.
-            Default: 0.0
-        proj_drop_rate (float, optional): Dropout ratio of output. Default: 0.
-        init_cfg (dict | None, optional): The Config for initialization.
-            Default: None.
-    """
-
-    def __init__(self,
-                 embed_dims,
-                 num_heads,
-                 window_size,
-                 qkv_bias=True,
-                 qk_scale=None,
-                 attn_drop_rate=0.,
-                 proj_drop_rate=0.,
-                 init_cfg=None):
-
-        super().__init__(init_cfg=init_cfg)
-        self.embed_dims = embed_dims
-        self.num_heads = num_heads
-        head_embed_dims = embed_dims // num_heads
-        self.scale = qk_scale or head_embed_dims**-0.5
-        if qkv_bias:
-            self.q_bias = nn.Parameter(torch.zeros(embed_dims))
-            self.v_bias = nn.Parameter(torch.zeros(embed_dims))
-        else:
-            self.q_bias = None
-            self.v_bias = None
-
-        if window_size:
-            self.window_size = window_size
-            self.num_relative_distance = (2 * window_size[0] -
-                                          1) * (2 * window_size[1] - 1) + 3
-            self.relative_position_bias_table = nn.Parameter(
-                torch.zeros(self.num_relative_distance,
-                            num_heads))  # 2*Wh-1 * 2*Ww-1, nH
-            # cls to token & token 2 cls & cls to cls
-
-            # get pair-wise relative position index for
-            # each token inside the window
-            coords_h = torch.arange(window_size[0])
-            coords_w = torch.arange(window_size[1])
-            coords = torch.stack(torch.meshgrid([coords_h,
-                                                 coords_w]))  # 2, Wh, Ww
-            coords_flatten = torch.flatten(coords, 1)  # 2, Wh*Ww
-            relative_coords = coords_flatten[:, :, None] - \
-                coords_flatten[:, None, :]
-            relative_coords = relative_coords.permute(
-                1, 2, 0).contiguous()  # Wh*Ww, Wh*Ww, 2
-            relative_coords[:, :,
-                            0] += window_size[0] - 1  # shift to start from 0
-            relative_coords[:, :, 1] += window_size[1] - 1
-            relative_coords[:, :, 0] *= 2 * window_size[1] - 1
-            relative_position_index = torch.zeros(
-                size=(window_size[0] * window_size[1] + 1, ) * 2,
-                dtype=relative_coords.dtype)
-            relative_position_index[1:, 1:] = relative_coords.sum(
-                -1)  # Wh*Ww, Wh*Ww
-            relative_position_index[0, 0:] = self.num_relative_distance - 3
-            relative_position_index[0:, 0] = self.num_relative_distance - 2
-            relative_position_index[0, 0] = self.num_relative_distance - 1
-
-            self.register_buffer('relative_position_index',
-                                 relative_position_index)
-
-        else:
-            self.window_size = None
-            self.relative_position_bias_table = None
-            self.relative_position_index = None
-
-        self.qkv = nn.Linear(embed_dims, embed_dims * 3, bias=False)
-        self.attn_drop = nn.Dropout(attn_drop_rate)
-        self.proj = nn.Linear(embed_dims, embed_dims)
-        self.proj_drop = nn.Dropout(proj_drop_rate)
-
-    def init_weights(self):
-        trunc_normal_(self.relative_position_bias_table, std=0.02)
-
-    def forward(self, x):
-        """
-        Args:
-
-            x (tensor): input features with shape of (num_windows*B, N, C)
-            mask (tensor | None, Optional): mask with shape of (num_windows,
-                Wh*Ww, Wh*Ww), value should be between (-inf, 0].
-        """
-        B, N, C = x.shape
-        qkv_bias = None
-        if self.q_bias is not None:
-            qkv_bias = torch.cat(
-                (self.q_bias,
-                 torch.zeros_like(self.v_bias,
-                                  requires_grad=False), self.v_bias))
-
-        qkv = F.linear(input=x, weight=self.qkv.weight, bias=qkv_bias)
-        qkv = qkv.reshape(B, N, 3, self.num_heads, -1).permute(2, 0, 3, 1, 4)
-        q, k, v = qkv[0], qkv[1], qkv[2]
-
-        q = q * self.scale
-        attn = (q @ k.transpose(-2, -1))
-
-        if self.relative_position_bias_table is not None:
-            relative_position_bias = \
-                self.relative_position_bias_table[
-                    self.relative_position_index.view(-1)].view(
-                    self.window_size[0] * self.window_size[1] + 1,
-                    self.window_size[0] * self.window_size[1] + 1, -1)
-            relative_position_bias = relative_position_bias.permute(
-                2, 0, 1).contiguous()  # nH, Wh*Ww, Wh*Ww
-            attn = attn + relative_position_bias.unsqueeze(0)
-
-        attn = attn.softmax(dim=-1)
-        attn = self.attn_drop(attn)
-
-        x = (attn @ v).transpose(1, 2).reshape(B, N, C)
-        x = self.proj(x)
-        x = self.proj_drop(x)
-        return x
 
 
 class TransformerEncoderLayer(BaseModule):
@@ -189,66 +53,33 @@ class TransformerEncoderLayer(BaseModule):
                  qkv_bias=True,
                  act_cfg=dict(type='GELU'),
                  norm_cfg=dict(type='LN'),
-                 batch_first=True,
-                 window_size=None,
-                 init_values=None):
+                 batch_first=True):
         super(TransformerEncoderLayer, self).__init__()
 
         self.norm1_name, norm1 = build_norm_layer(
             norm_cfg, embed_dims, postfix=1)
         self.add_module(self.norm1_name, norm1)
-        if window_size is not None:
-            self.attn = BEiTAttention(
-                embed_dims=embed_dims,
-                num_heads=num_heads,
-                window_size=window_size,
-                qkv_bias=True,
-                qk_scale=None,
-                attn_drop_rate=0.,
-                proj_drop_rate=0.,
-                init_cfg=None)
-            self.ffn = FFN(
-                embed_dims=embed_dims,
-                feedforward_channels=feedforward_channels,
-                num_fcs=num_fcs,
-                ffn_drop=0.,
-                dropout_layer=None,
-                act_cfg=act_cfg,
-                add_identity=False)
-        else:
-            self.attn = MultiheadAttention(
-                embed_dims=embed_dims,
-                num_heads=num_heads,
-                attn_drop=attn_drop_rate,
-                proj_drop=drop_rate,
-                dropout_layer=dict(type='DropPath', drop_prob=drop_path_rate),
-                batch_first=batch_first,
-                bias=qkv_bias)
-            self.ffn = FFN(
-                embed_dims=embed_dims,
-                feedforward_channels=feedforward_channels,
-                num_fcs=num_fcs,
-                ffn_drop=drop_rate,
-                dropout_layer=dict(type='DropPath', drop_prob=drop_path_rate),
-                act_cfg=act_cfg)
+
+        self.attn = MultiheadAttention(
+            embed_dims=embed_dims,
+            num_heads=num_heads,
+            attn_drop=attn_drop_rate,
+            proj_drop=drop_rate,
+            dropout_layer=dict(type='DropPath', drop_prob=drop_path_rate),
+            batch_first=batch_first,
+            bias=qkv_bias)
 
         self.norm2_name, norm2 = build_norm_layer(
             norm_cfg, embed_dims, postfix=2)
         self.add_module(self.norm2_name, norm2)
 
-        # NOTE: drop path for stochastic depth,
-        # we shall see if this is better than dropout here
-        dropout_layer = dict(type='DropPath', drop_prob=drop_path_rate)
-        self.drop_path = build_dropout(
-            dropout_layer) if dropout_layer else nn.Identity()
-
-        if init_values is not None:
-            self.gamma_1 = nn.Parameter(
-                init_values * torch.ones((embed_dims)), requires_grad=True)
-            self.gamma_2 = nn.Parameter(
-                init_values * torch.ones((embed_dims)), requires_grad=True)
-        else:
-            self.gamma_1, self.gamma_2 = None, None
+        self.ffn = FFN(
+            embed_dims=embed_dims,
+            feedforward_channels=feedforward_channels,
+            num_fcs=num_fcs,
+            ffn_drop=drop_rate,
+            dropout_layer=dict(type='DropPath', drop_prob=drop_path_rate),
+            act_cfg=act_cfg)
 
     @property
     def norm1(self):
@@ -259,44 +90,9 @@ class TransformerEncoderLayer(BaseModule):
         return getattr(self, self.norm2_name)
 
     def forward(self, x):
-        if self.gamma_1 is None:
-            x = self.attn(self.norm1(x), identity=x)
-            x = self.ffn(self.norm2(x), identity=x)
-        else:
-            x = x + self.drop_path(self.gamma_1 * self.attn(self.norm1(x)))
-            x = x + self.drop_path(self.gamma_2 * self.ffn(self.norm2(x)))
+        x = self.attn(self.norm1(x), identity=x)
+        x = self.ffn(self.norm2(x), identity=x)
         return x
-
-
-class BEiTPatchEmbed(nn.Module):
-    """Image to Patch Embedding."""
-
-    def __init__(self, img_size=224, patch_size=16, in_chans=3, embed_dim=768):
-        super().__init__()
-        img_size = to_2tuple(img_size)
-        patch_size = to_2tuple(patch_size)
-        num_patches = (img_size[1] // patch_size[1]) * (
-            img_size[0] // patch_size[0])
-        self.patch_shape = (img_size[0] // patch_size[0],
-                            img_size[1] // patch_size[1])
-        self.img_size = img_size
-        self.patch_size = patch_size
-        self.num_patches = num_patches
-
-        self.proj = nn.Conv2d(
-            in_chans, embed_dim, kernel_size=patch_size, stride=patch_size)
-
-    def forward(self, x, **kwargs):
-        B, C, H, W = x.shape
-        # FIXME look at relaxing size constraints
-        # assert H == self.img_size[0] and W == self.img_size[1], \
-        #     f"Input image size ({H}*{W}) doesn't match model
-        # ({self.img_size[0]}*{self.img_size[1]})."
-        x = self.proj(x)
-        Hp, Wp = x.shape[2], x.shape[3]
-
-        x = x.flatten(2).transpose(1, 2)
-        return x, (Hp, Wp)
 
 
 @BACKBONES.register_module()
@@ -306,7 +102,6 @@ class VisionTransformer(BaseModule):
     This backbone is the implementation of `An Image is Worth 16x16 Words:
     Transformers for Image Recognition at
     Scale <https://arxiv.org/abs/2010.11929>`_.
-
     Args:
         img_size (int | tuple): Input image size. Default: 224.
         patch_size (int): The patch size. Default: 16.
@@ -346,9 +141,6 @@ class VisionTransformer(BaseModule):
         with_cp (bool): Use checkpoint or not. Using checkpoint will save
             some memory while slowing down the training speed. Default: False.
         pretrained (str, optional): model pretrained path. Default: None.
-        use_rel_pos_bias (bool): Whether to use BEiTAttention
-        init_values (float): Initialize the values of BEiTAttention and FFN
-        with learnable scaling.
         init_cfg (dict or list[dict], optional): Initialization config dict.
             Default: None.
     """
@@ -377,8 +169,6 @@ class VisionTransformer(BaseModule):
                  norm_eval=False,
                  with_cp=False,
                  pretrained=None,
-                 use_rel_pos_bias=False,
-                 init_values=None,
                  init_cfg=None):
         super(VisionTransformer, self).__init__(init_cfg=init_cfg)
 
@@ -411,36 +201,25 @@ class VisionTransformer(BaseModule):
         self.with_cp = with_cp
         self.pretrained = pretrained
 
-        if use_rel_pos_bias:
-            self.patch_embed = BEiTPatchEmbed(
-                img_size=img_size[0],
-                patch_size=patch_size,
-                in_chans=in_channels,
-                embed_dim=embed_dims)
-        else:
-            self.patch_embed = PatchEmbed(
-                in_channels=in_channels,
-                embed_dims=embed_dims,
-                conv_type='Conv2d',
-                kernel_size=patch_size,
-                stride=patch_size,
-                padding='corner',
-                norm_cfg=norm_cfg if patch_norm else None,
-                init_cfg=None,
-            )
+        self.patch_embed = PatchEmbed(
+            in_channels=in_channels,
+            embed_dims=embed_dims,
+            conv_type='Conv2d',
+            kernel_size=patch_size,
+            stride=patch_size,
+            padding='corner',
+            norm_cfg=norm_cfg if patch_norm else None,
+            init_cfg=None,
+        )
 
-        self.use_rel_pos_bias = use_rel_pos_bias
         num_patches = (img_size[0] // patch_size) * \
             (img_size[1] // patch_size)
 
-        window_size = (img_size[0] // patch_size, img_size[1] // patch_size)
-        self.patch_shape = window_size
         self.with_cls_token = with_cls_token
         self.output_cls_token = output_cls_token
         self.cls_token = nn.Parameter(torch.zeros(1, 1, embed_dims))
-        if not use_rel_pos_bias:
-            self.pos_embed = nn.Parameter(
-                torch.zeros(1, num_patches + 1, embed_dims))
+        self.pos_embed = nn.Parameter(
+            torch.zeros(1, num_patches + 1, embed_dims))
         self.drop_after_pos = nn.Dropout(p=drop_rate)
 
         if isinstance(out_indices, int):
@@ -470,9 +249,7 @@ class VisionTransformer(BaseModule):
                     qkv_bias=qkv_bias,
                     act_cfg=act_cfg,
                     norm_cfg=norm_cfg,
-                    batch_first=True,
-                    window_size=window_size if use_rel_pos_bias else None,
-                    init_values=init_values))
+                    batch_first=True))
 
         self.final_norm = final_norm
         if final_norm:
@@ -480,35 +257,11 @@ class VisionTransformer(BaseModule):
                 norm_cfg, embed_dims, postfix=1)
             self.add_module(self.norm1_name, norm1)
 
-        if use_rel_pos_bias:
-            self.fix_init_weight()
-
     @property
     def norm1(self):
         return getattr(self, self.norm1_name)
 
-    def fix_init_weight(self):
-
-        def rescale(param, layer_id):
-            param.div_(math.sqrt(2.0 * layer_id))
-
-        for layer_id, layer in enumerate(self.layers):
-            rescale(layer.attn.proj.weight.data, layer_id + 1)
-            rescale(layer.ffn.layers[1].weight.data, layer_id + 1)
-
     def init_weights(self):
-
-        def _init_weights(m):
-            if isinstance(m, nn.Linear):
-                trunc_normal_(m.weight, std=.02)
-                if isinstance(m, nn.Linear) and m.bias is not None:
-                    nn.init.constant_(m.bias, 0)
-            elif isinstance(m, nn.LayerNorm):
-                nn.init.constant_(m.bias, 0)
-                nn.init.constant_(m.weight, 1.0)
-
-        self.apply(_init_weights)
-
         if (isinstance(self.init_cfg, dict)
                 and self.init_cfg.get('type') == 'Pretrained'):
             logger = get_root_logger()
@@ -532,98 +285,6 @@ class VisionTransformer(BaseModule):
                         state_dict['pos_embed'],
                         (h // self.patch_size, w // self.patch_size),
                         (pos_size, pos_size), self.interpolate_mode)
-
-            elif self.use_rel_pos_bias:
-                all_keys = list(state_dict.keys())
-                for key in all_keys:
-                    if 'relative_position_index' in key:
-                        state_dict.pop(key)
-
-                    if 'relative_position_bias_table' in key:
-                        rel_pos_bias = state_dict[key]
-                        src_num_pos, num_attn_heads = rel_pos_bias.size()
-                        dst_num_pos, _ = self.state_dict()[key].size()
-                        dst_patch_shape = self.patch_shape
-                        if dst_patch_shape[0] != dst_patch_shape[1]:
-                            raise NotImplementedError()
-                        num_extra_tokens = dst_num_pos - (
-                            dst_patch_shape[0] * 2 - 1) * (
-                                dst_patch_shape[1] * 2 - 1)
-                        src_size = int((src_num_pos - num_extra_tokens)**0.5)
-                        dst_size = int((dst_num_pos - num_extra_tokens)**0.5)
-                        if src_size != dst_size:
-                            extra_tokens = rel_pos_bias[-num_extra_tokens:, :]
-                            rel_pos_bias = rel_pos_bias[:-num_extra_tokens, :]
-
-                            def geometric_progression(a, r, n):
-                                return a * (1.0 - r**n) / (1.0 - r)
-
-                            left, right = 1.01, 1.5
-                            while right - left > 1e-6:
-                                q = (left + right) / 2.0
-                                gp = geometric_progression(1, q, src_size // 2)
-                                if gp > dst_size // 2:
-                                    right = q
-                                else:
-                                    left = q
-
-                            # if q > 1.13492:
-                            #     q = 1.13492
-
-                            dis = []
-                            cur = 1
-                            for i in range(src_size // 2):
-                                dis.append(cur)
-                                cur += q**(i + 1)
-
-                            r_ids = [-_ for _ in reversed(dis)]
-
-                            x = r_ids + [0] + dis
-                            y = r_ids + [0] + dis
-
-                            t = dst_size // 2.0
-                            dx = np.arange(-t, t + 0.1, 1.0)
-                            dy = np.arange(-t, t + 0.1, 1.0)
-
-                            all_rel_pos_bias = []
-
-                            for i in range(num_attn_heads):
-                                z = rel_pos_bias[:, i].view(
-                                    src_size, src_size).float().numpy()
-                                f = interpolate.interp2d(x, y, z, kind='cubic')
-                                all_rel_pos_bias.append(
-                                    torch.Tensor(f(dx, dy)).contiguous().view(
-                                        -1, 1).to(rel_pos_bias.device))
-
-                            rel_pos_bias = torch.cat(all_rel_pos_bias, dim=-1)
-                            new_rel_pos_bias = torch.cat(
-                                (rel_pos_bias, extra_tokens), dim=0)
-                            state_dict[key] = new_rel_pos_bias
-
-                # interpolate position bias table if needed
-                relative_position_bias_table_keys = [
-                    k for k in state_dict.keys()
-                    if 'relative_position_bias_table' in k
-                ]
-                for table_key in relative_position_bias_table_keys:
-                    table_pretrained = state_dict[table_key]
-                    table_current = self.state_dict()[table_key]
-                    L1, nH1 = table_pretrained.size()
-                    L2, nH2 = table_current.size()
-                    if nH1 != nH2:
-                        logger.warning(f'Error in loading {table_key}, pass')
-                    else:
-                        if L1 != L2:
-                            S1 = int(L1**0.5)
-                            S2 = int(L2**0.5)
-                            table_pretrained_resized = F.interpolate(
-                                table_pretrained.permute(1, 0).view(
-                                    1, nH1, S1, S1),
-                                size=(S2, S2),
-                                mode='bicubic')
-                            state_dict[
-                                table_key] = table_pretrained_resized.view(
-                                    nH2, L2).permute(1, 0)
 
             self.load_state_dict(state_dict, False)
         elif self.init_cfg is not None:
@@ -715,8 +376,7 @@ class VisionTransformer(BaseModule):
         # stole cls_tokens impl from Phil Wang, thanks
         cls_tokens = self.cls_token.expand(B, -1, -1)
         x = torch.cat((cls_tokens, x), dim=1)
-        if not self.use_rel_pos_bias:
-            x = self._pos_embeding(x, hw_shape, self.pos_embed)
+        x = self._pos_embeding(x, hw_shape, self.pos_embed)
 
         if not self.with_cls_token:
             # Remove class token for transformer encoder input
